+++ pkg/meta/CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 1.18.0-dev
 
 - The `@redeclare` annotation is no longer considered experimental.
+- Two new TargetKinds are introduced: `TargetKind.exportDirective`, which
+  indicates an annotation is valid on an export directive, and
+  `TargetKind.partOfDirective`, which indicates an annnotation is valid on a
+  "part of" directive.
+- The `TargetKind.directive` value is deprecated in favor of the above new
+  TargetKinds and the existing `TargetKind.library`.
 
 ## 1.17.0